Separate BEP parsing and data
This is a step towards having a simpler and faster implementation for
the needs of the query sync that does not support sharding.

Bug: 327638725
Test: n/a (existing)

AOSP: 300518bc59dc4d72c797a07bbd786d779b5d68ac
