+++ base/src/com/google/idea/blaze/base/command/buildresult/BuildResultHelperBep.java
@@ -15,7 +15,7 @@
  */
 package com.google.idea.blaze.base.command.buildresult;
 
-import static com.google.idea.blaze.base.command.buildresult.bepparser.ParsedBepOutput.parseBepArtifacts;
+import static com.google.idea.blaze.base.command.buildresult.bepparser.BepParser.parseBepArtifacts;
 
 import com.google.idea.blaze.base.command.buildresult.bepparser.BuildEventStreamProvider;
 import com.intellij.openapi.diagnostic.Logger;
